getting referrer acquistion data from a query string implemented using a query string bindable - twit twoo!
